+++ esphome/components/qr_code/qr_code.cpp
@@ -28,13 +28,16 @@ void QrCode::set_ecc(qrcodegen_Ecc ecc) {
 void QrCode::generate_qr_code() {
   ESP_LOGV(TAG, "Generating QR code");
 
-  // Calculate buffer size needed to encode text for the QR code
-  // and allocate stack or heap memory for tempbuffer
-  size_t textLen = strlen(this->value_.c_str());
-  size_t buffer_length = qrcodegen_calcSegmentBufferSize(qrcodegen_Mode_ALPHANUMERIC, textLen);
-
-  SmallBufferWithHeapFallback<1024> buffer_alloc;  // Stack for small QR, heap for large
-  uint8_t *tempbuffer = buffer_alloc.get(buffer_length);
+#ifdef USE_ESP32
+  // ESP32 has 8KB stack, safe to allocate ~4KB buffer on stack
+  uint8_t tempbuffer[qrcodegen_BUFFER_LEN_MAX];
+#else
+  // Other platforms (ESP8266: 4KB, RP2040: 2KB, LibreTiny: ~4KB) have smaller stacks
+  // Use stack for small QR codes, heap for large ones to avoid stack overflow
+  size_t buffer_len = qrcodegen_calcSegmentBufferSize(qrcodegen_Mode_ALPHANUMERIC, this->value_.length());
+  SmallBufferWithHeapFallback<1024> buffer_alloc;
+  uint8_t *tempbuffer = buffer_alloc.get(buffer_len);
+#endif
 
   if (!qrcodegen_encodeText(this->value_.c_str(), tempbuffer, this->qr_, this->ecc_, qrcodegen_VERSION_MIN,
                             qrcodegen_VERSION_MAX, qrcodegen_Mask_AUTO, true)) {
